merge upstream

+++ Tests/AlgorithmRunner.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -17,7 +17,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Threading.Tasks;
 using NodaTime;
 using NUnit.Framework;
@@ -26,14 +25,15 @@ using QuantConnect.Configuration;
 using QuantConnect.Data;
 using QuantConnect.Interfaces;
 using QuantConnect.Lean.Engine;
-using QuantConnect.Lean.Engine.Alphas;
+using QuantConnect.Lean.Engine.DataFeeds;
 using QuantConnect.Lean.Engine.HistoricalData;
 using QuantConnect.Lean.Engine.Results;
 using QuantConnect.Lean.Engine.Setup;
+using QuantConnect.Lean.Engine.Storage;
 using QuantConnect.Logging;
+using QuantConnect.Orders;
 using QuantConnect.Packets;
 using QuantConnect.Securities;
-using QuantConnect.Tests.Common.Securities;
 using QuantConnect.Util;
 using HistoryRequest = QuantConnect.Data.HistoryRequest;
 
@@ -47,26 +47,38 @@ namespace QuantConnect.Tests
         public static AlgorithmRunnerResults RunLocalBacktest(
             string algorithm,
             Dictionary<string, string> expectedStatistics,
-            AlphaRuntimeStatistics expectedAlphaStatistics,
             Language language,
             AlgorithmStatus expectedFinalStatus,
             DateTime? startDate = null,
             DateTime? endDate = null,
             string setupHandler = "RegressionSetupHandlerWrapper",
-            decimal? initialCash = null)
+            decimal? initialCash = null,
+            string algorithmLocation = null,
+            bool returnLogs = false)
         {
             AlgorithmManager algorithmManager = null;
             var statistics = new Dictionary<string, string>();
-            var alphaStatistics = new AlphaRuntimeStatistics(new TestAccountCurrencyProvider());
             BacktestingResultHandler results = null;
 
             Composer.Instance.Reset();
             SymbolCache.Clear();
+            TextSubscriptionDataSourceReader.ClearCache();
+            MarketOnCloseOrder.SubmissionTimeBuffer = MarketOnCloseOrder.DefaultSubmissionTimeBuffer;
+
+            // clean up object storage
+            if (Directory.Exists(LocalObjectStore.DefaultObjectStore))
+            {
+                Directory.Delete(LocalObjectStore.DefaultObjectStore, true);
+            }
 
             var ordersLogFile = string.Empty;
             var logFile = $"./regression/{algorithm}.{language.ToLower()}.log";
             Directory.CreateDirectory(Path.GetDirectoryName(logFile));
             File.Delete(logFile);
+            var logs = new List<string>();
+
+            var reducedDiskSize = TestContext.Parameters.Exists("reduced-disk-size") &&
+                bool.Parse(TestContext.Parameters["reduced-disk-size"]);
 
             try
             {
@@ -74,36 +86,53 @@ namespace QuantConnect.Tests
                 Config.Set("algorithm-type-name", algorithm);
                 Config.Set("live-mode", "false");
                 Config.Set("environment", "");
-                Config.Set("messaging-handler", "QuantConnect.Messaging.Messaging");
+                Config.Set("messaging-handler", "QuantConnect.Tests.RegressionTestMessageHandler");
                 Config.Set("job-queue-handler", "QuantConnect.Queues.JobQueue");
                 Config.Set("setup-handler", setupHandler);
                 Config.Set("history-provider", "RegressionHistoryProviderWrapper");
                 Config.Set("api-handler", "QuantConnect.Api.Api");
                 Config.Set("result-handler", "QuantConnect.Lean.Engine.Results.RegressionResultHandler");
                 Config.Set("algorithm-language", language.ToString());
-                Config.Set("algorithm-location",
-                    language == Language.Python
-                        ? "../../../Algorithm.Python/" + algorithm + ".py"
-                        : "QuantConnect.Algorithm." + language + ".dll");
+                if (string.IsNullOrEmpty(algorithmLocation))
+                {
+                    Config.Set("algorithm-location",
+                        language == Language.Python
+                            ? "../../../Algorithm.Python/" + algorithm + ".py"
+                            : "QuantConnect.Algorithm." + language + ".dll");
+                }
+                else
+                {
+                    Config.Set("algorithm-location", algorithmLocation);
+                }
 
+                // Store initial log variables
+                var initialLogHandler = Log.LogHandler;
+                var initialDebugEnabled = Log.DebuggingEnabled;
 
-                var debugEnabled = Log.DebuggingEnabled;
+                var newLogHandlers = new List<ILogHandler>() { MaintainLogHandlerAttribute.LogHandler };
+                // Use our current test LogHandler and a FileLogHandler
+                if (!reducedDiskSize)
+                {
+                    newLogHandlers.Add(new FileLogHandler(logFile, false));
+                }
+                if (returnLogs)
+                {
+                    var storeLog = (string logMessage) => logs.Add(logMessage);
+                    newLogHandlers.Add(new FunctionalLogHandler(storeLog, storeLog, storeLog));
+                }
 
-
-                var logHandlers = new ILogHandler[] {new ConsoleLogHandler(), new FileLogHandler(logFile, false)};
-                using (Log.LogHandler = new CompositeLogHandler(logHandlers))
+                using (Log.LogHandler = new CompositeLogHandler(newLogHandlers.ToArray()))
                 using (var algorithmHandlers = LeanEngineAlgorithmHandlers.FromConfiguration(Composer.Instance))
                 using (var systemHandlers = LeanEngineSystemHandlers.FromConfiguration(Composer.Instance))
                 using (var workerThread  = new TestWorkerThread())
                 {
-                    Log.DebuggingEnabled = true;
+                    Log.DebuggingEnabled = !reducedDiskSize;
 
                     Log.Trace("");
                     Log.Trace("{0}: Running " + algorithm + "...", DateTime.UtcNow);
                     Log.Trace("");
 
                     // run the algorithm in its own thread
-
                     var engine = new Lean.Engine.Engine(systemHandlers, algorithmHandlers, false);
                     Task.Factory.StartNew(() =>
                     {
@@ -120,6 +149,12 @@ namespace QuantConnect.Tests
                             }
                             algorithmManager = new AlgorithmManager(false, job);
 
+                            var regressionTestMessageHandler = systemHandlers.Notify as RegressionTestMessageHandler;
+                            if (regressionTestMessageHandler != null)
+                            {
+                                regressionTestMessageHandler.SetAlgorithmManager(algorithmManager);
+                            }
+
                             systemHandlers.LeanManager.Initialize(systemHandlers, algorithmHandlers, job, algorithmManager);
 
                             engine.Run(job, algorithmManager, algorithmPath, workerThread);
@@ -131,15 +166,19 @@ namespace QuantConnect.Tests
                         }
                     }).Wait();
 
-                    var backtestingResultHandler = (BacktestingResultHandler)algorithmHandlers.Results;
-                    results = backtestingResultHandler;
-                    statistics = backtestingResultHandler.FinalStatistics;
+                    var regressionResultHandler = (RegressionResultHandler)algorithmHandlers.Results;
+                    results = regressionResultHandler;
+                    statistics = regressionResultHandler.FinalStatistics;
 
-                    var defaultAlphaHandler = (DefaultAlphaHandler) algorithmHandlers.Alphas;
-                    alphaStatistics = defaultAlphaHandler.RuntimeStatistics;
-
-                    Log.DebuggingEnabled = debugEnabled;
+                    if (expectedFinalStatus == AlgorithmStatus.Completed && regressionResultHandler.HasRuntimeError)
+                    {
+                        Assert.Fail($"There was a runtime error running the algorithm");
+                    }
                 }
+
+                // Reset settings to initial values
+                Log.LogHandler = initialLogHandler;
+                Log.DebuggingEnabled = initialDebugEnabled;
             }
             catch (Exception ex)
             {
@@ -154,63 +193,48 @@ namespace QuantConnect.Tests
                 Assert.Fail($"Algorithm state should be {expectedFinalStatus} and is: {algorithmManager?.State}");
             }
 
-            foreach (var stat in expectedStatistics)
+            foreach (var expectedStat in expectedStatistics)
             {
-                Assert.AreEqual(true, statistics.ContainsKey(stat.Key), "Missing key: " + stat.Key);
-                Assert.AreEqual(stat.Value, statistics[stat.Key], "Failed on " + stat.Key);
+                string result;
+                Assert.IsTrue(statistics.TryGetValue(expectedStat.Key, out result), "Missing key: " + expectedStat.Key);
+
+                // normalize -0 & 0, they are the same thing
+                var expected = expectedStat.Value;
+                if (expected == "-0")
+                {
+                    expected = "0";
+                }
+
+                if (result == "-0")
+                {
+                    result = "0";
+                }
+
+                Assert.AreEqual(expected, result, "Failed on " + expectedStat.Key);
             }
 
-            if (expectedAlphaStatistics != null)
+            if (!reducedDiskSize)
             {
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.MeanPopulationScore.Direction);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.MeanPopulationScore.Magnitude);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.RollingAveragedPopulationScore.Direction);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.RollingAveragedPopulationScore.Magnitude);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.LongShortRatio);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.TotalInsightsClosed);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.TotalInsightsGenerated);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.TotalAccumulatedEstimatedAlphaValue);
-                AssertAlphaStatistics(expectedAlphaStatistics, alphaStatistics, s => s.TotalInsightsAnalysisCompleted);
+                // we successfully passed the regression test, copy the log file so we don't have to continually
+                // re-run master in order to compare against a passing run
+                var passedFile = logFile.Replace("./regression/", "./passed/");
+                Directory.CreateDirectory(Path.GetDirectoryName(passedFile));
+                File.Delete(passedFile);
+                File.Copy(logFile, passedFile);
+
+                var passedOrderLogFile = ordersLogFile.Replace("./regression/", "./passed/");
+                Directory.CreateDirectory(Path.GetDirectoryName(passedFile));
+                File.Delete(passedOrderLogFile);
+                if (File.Exists(ordersLogFile)) File.Copy(ordersLogFile, passedOrderLogFile);
+
             }
-
-            // we successfully passed the regression test, copy the log file so we don't have to continually
-            // re-run master in order to compare against a passing run
-            var passedFile = logFile.Replace("./regression/", "./passed/");
-            Directory.CreateDirectory(Path.GetDirectoryName(passedFile));
-            File.Delete(passedFile);
-            File.Copy(logFile, passedFile);
-
-            var passedOrderLogFile = ordersLogFile.Replace("./regression/", "./passed/");
-            Directory.CreateDirectory(Path.GetDirectoryName(passedFile));
-            File.Delete(passedOrderLogFile);
-            if (File.Exists(ordersLogFile)) File.Copy(ordersLogFile, passedOrderLogFile);
-
-            return new AlgorithmRunnerResults(algorithm, language, algorithmManager, results);
-        }
-
-        private static void AssertAlphaStatistics(AlphaRuntimeStatistics expected, AlphaRuntimeStatistics actual, Expression<Func<AlphaRuntimeStatistics, object>> selector)
-        {
-            // extract field name from expression
-            var field = selector.AsEnumerable().OfType<MemberExpression>().First().ToString();
-            field = field.Substring(field.IndexOf('.') + 1);
-
-            var func = selector.Compile();
-            var expectedValue = func(expected);
-            var actualValue = func(actual);
-            if (expectedValue is double)
-            {
-                Assert.AreEqual((double)expectedValue, (double)actualValue, 1e-4, "Failed on alpha statistics " + field);
-            }
-            else
-            {
-                Assert.AreEqual(expectedValue, actualValue, "Failed on alpha statistics " + field);
-            }
+            return new AlgorithmRunnerResults(algorithm, language, algorithmManager, results, logs);
         }
 
         /// <summary>
         /// Used to intercept the algorithm instance to aid the <see cref="RegressionHistoryProviderWrapper"/>
         /// </summary>
-        internal class RegressionSetupHandlerWrapper : BacktestingSetupHandler
+        public class RegressionSetupHandlerWrapper : BacktestingSetupHandler
         {
             public static IAlgorithm Algorithm { get; protected set; }
             public override IAlgorithm CreateAlgorithmInstance(AlgorithmNodePacket algorithmNodePacket, string assemblyPath)
@@ -228,12 +252,13 @@ namespace QuantConnect.Tests
         /// <summary>
         /// Used to perform checks against history requests for all regression algorithms
         /// </summary>
-        class RegressionHistoryProviderWrapper : SubscriptionDataReaderHistoryProvider
+        public class RegressionHistoryProviderWrapper : SubscriptionDataReaderHistoryProvider
         {
             public override IEnumerable<Slice> GetHistory(IEnumerable<HistoryRequest> requests, DateTimeZone sliceTimeZone)
             {
                 requests = requests.ToList();
-                if (requests.Any(r => RegressionSetupHandlerWrapper.Algorithm.UniverseManager.ContainsKey(r.Symbol)))
+                if (requests.Any(r => RegressionSetupHandlerWrapper.Algorithm.UniverseManager.ContainsKey(r.Symbol)
+                    && (r.Symbol.SecurityType != SecurityType.Future || !r.Symbol.IsCanonical())))
                 {
                     throw new Exception("History requests should not be submitted for universe symbols");
                 }
@@ -241,7 +266,7 @@ namespace QuantConnect.Tests
             }
         }
 
-        class TestWorkerThread : WorkerThread
+        public class TestWorkerThread : WorkerThread
         {
         }
     }
